Divide into two separate components internally
Since the selects in single mode and multiple mode have many differences
(subtle ones, but many) the code was more complicated than what it had
to be.

Now there is 4 components:

- ember-power-select/single: Handles single selection only.
- ember-power-select/multiple: Handles multiple selection.
- ember-power-select/base: Abstract. Superclass of the former 2.
- ember-power-select: Proxy component that internally uses the single or
  multiple version depending if it receives `multiple=true` or not.

Quite happy with the result. Each component is under 80 lines and
the logic is much easier to follow. With less exceptions.

Closes #23
