+++ app/components/ember-power-select/multiple.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-power-select/components/ember-power-select/multiple';
