+++ app/components/ember-power-select/single.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-power-select/components/ember-power-select/single';
