+++ tests/dummy/app/templates/-selected-country.hbs
@@ -1,2 +1,2 @@
-<img src="{{selectedOption.flagUrl}}" class="icon-flag">
-{{selectedOption.name}}
+<img src="{{selection.flagUrl}}" class="icon-flag">
+{{selection.name}}
